sunxi-hdmi: Speed up hdmi initialization
Before this patch we started in this useless "idle" state, from where we
would set the next state to be HDMI_State_Wait_Hpd and then return from
hdmi_main_task_loop to Hdmi_run_thread, which would then sleep 2 seconds
before checking if we've a link. So in essence the hdmi code would just
start with sleeping for 2 seconds for no good reason.

This patch gets rid of the idle state, shaving of 2 seconds of hdmi
init time. It also speeds up Hdmi_run_thread when we're not in a stable
state, so that if we're waiting for video-config (which we will get
pretty quickly after being brought up), the config arriving will get
noticed quickly.

This changes the hdmi init time from taking seconds to taking 10s of ms.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>