HACK: sunxi: video: change kernel framebuffer mapping to writecombine
The kernelspace mapping of framebuffer (needed for fbcon) had CPU
caching enabled by some mistake. This caused glitches and artefacts
in the framebuffer console, which can be easily reproduced as reported
on IRC:

  <buZz> supereasy to reproduce;
         1) type a multiword command on (fbdev) console
         2) ctrl-left a word
         3) delete its first character
         4) tada, sticky pixels

The solution is to change the kernel framebuffer memory mapping to
writecombine, just like it is done for the userspace mapping.

This patch is definitely *not* a clean solution. It is cutting some
corners by doing exactly what was NAK'ed by the ARM kernel maintainers:
    http://lists.infradead.org/pipermail/linux-arm-kernel/2011-April/048844.html
    http://lists.infradead.org/pipermail/linux-arm-kernel/2011-June/052233.html

But the correct solution seems to be rather invasive, and I abandoned
my attempt at implementing it. In the end, we may want to switch to
CMA and/or dma_alloc_writecombine() for all these memory buffers
and drop the legacy code.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>
Reviewed-By: Luc Verhaegen <libv@skynet.be>