sunxi-disp: Use mode names like pal and ntsc for specifying tv-out modes
Some of the tv modes have the exact same dimensions but are different in
things like how the color is encoded, so specifying <width>x<height> is not
enough to uniquely identify them.

Therefore this patch changes the screen0_output_mode value to be a mode-name
like pal, ntsc or pal-nc when the screen0_output_type is set to tv.

As an added bonus this patch also limits which modes from the tv_mode table
can be selected, so that the user can no longer select hdmi only modes when
using tv-out.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>