Drop sun3i from stage/sunxi-3.4
This patchset drops the sun3i architecture from the kernel tree. The
main reason is that it is unsupported, unmaintained and untested. It is
in a bitrotten state and nobody uses it anymore.

If anybody wants to step up and clean up, unify and maintain sun3i, this
is your last chance!

Signed-off-by: Oliver Schinagl <oliver@schinagl.nl>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>