sunxi-hci: Remove code believed to be no-op
USBC_Phy_GetCsr returns same address for the both
controllers 1 and 2 and it actually returns the
address of the first controller. That is why the
UsbInitPhy alway changes some properties of the musb
controller. Also the exact same UsbInitPhy exists in
drivers/usb/sunxi_usb/usbc/usbc_phy.c.
It seems like it was copy pasted to ehci/ohci without a
reason. After a lot of experimenting and testing it appears
the UsbInitPhy can be dropped safely along with its satellites.
The testing was performed on Mele A1000 (sun4i).

Signed-off-by: Arokux <arokux@gmail.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>