sunxi-emac: Use physical rather then virtual address for rx by dma
After the port to the dma_compat.h code sunxi_dma_enqueue should be
passed a physical buffer address, but this was overseen.

This did not cause any problem in practice because:
1) DMA_CPU_TRRESHOLD is set to 2000, and since ethernet's MTU is 1518
bytes this is never reached, so the dma rx code path is never used
2) Even when the DMA_CPU_TRRESHOLD is lowered, the virt address will work too
since the only difference is bit 31 of the address and that seems to be
ignored by the address decoder

Still this ough to be fixed and that is what this patch does, while at it
also switch to using dma_map_single / dma_unmap_single rather then DIY
cache flushing.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>