sun4i-keyboard: Make loading controlled through fex
Wether or not a board actually has the evaluation board style tablet
buttons hooked up to adc0 through a resistor network is something which
cannot be autodetected and thus typically belongs in the fex file.

Note this patch makes the new behavior of not registering the input-dev unless
there is:

[tabletkeys_para]
tabletkeys_used = 1

In the fex-file configurable from Kconfig and default its to off, for
compatility with old setups / fex files.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>