phjanderson (freaktab) video fix
In rk3188_lcdc.c is a handler for vsync interrupts:
   static irqreturn_t rk3188_lcdc_isr(intirq,void*dev_id)
Basically this code triggers all the listeners for vsync events
(vsync_info.wait). rk_fb.c hooks into this event:
   static int rk_fb_wait_for_vsync_thread(void*data)
Basically this function starts a thread that listens to the vsync events
and publishes those to the system (sysfs_notify).
Now look at this particular line (this commit's diff)

It says: wait for the vsync event, but the condition needs to be
!ktime_equal(timestamp, dev_drv->vsync_info.timestamp) && dev_drv-
>vsync_info.active. The last bit is to see if the vsync thread is
running, the first part however... no clue. Seems to me some kind of
workaround to prevent the vsync event to be accidentally triggered twice
for the same vsync. I removed !ktime_equal(timestamp, dev_drv-
>vsync_info.timestamp) && part from that line and tried to play a video:
perfectly smooth!

Reference:
http://www.freaktab.com/showthread.php?5939-Fixing-RK3188-video-playback&p=81489&viewfull=1#post81489